requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via pip-tools
django==1.11.2
djangorestframework==3.6.3
first==2.0.1              # via pip-tools
pip-tools==1.9.0
py==1.4.34                # via pytest
pytest-runner==2.11.1
pytest==3.1.2
pytz==2017.2              # via django
six==1.10.0               # via pip-tools
wheel==0.29.0
